Split sampler.go into multiple files
