+++ pkg/sampler/sampler.go
@@ -2,10 +2,6 @@ package sampler
 
 import (
 	"bufio"
-	"crypto/md5"
-	"crypto/tls"
-	"encoding/hex"
-	"encoding/json"
 	"fmt"
 	"net"
 	"net/http"
@@ -14,25 +10,6 @@ import (
 	"strings"
 	"time"
 )
-
-type Target struct {
-	URL      string
-	Name     string
-	Interval int
-	// metadata
-	Tags               []string
-	Attributes         map[string]string
-	Hash               string
-	RequestHeaders     map[string]string
-	InsecureSkipVerify bool
-}
-
-func (t *Target) SetHash() {
-	jsonTarget, _ := json.Marshal(t)
-	hasher := md5.New()
-	hasher.Write(jsonTarget)
-	t.Hash = hex.EncodeToString(hasher.Sum(nil))
-}
 
 type Sample struct {
 	StatusCode      int
@@ -127,29 +104,6 @@ func Ping(target Target, timeout int) (sample Sample, err error) {
 	return
 }
 
-func dial(t Target) (net.Conn, error) {
-	u, err := url.Parse(t.URL)
-	if err != nil {
-		return nil, err
-	}
-
-	host, err := hostString(u)
-	if err != nil {
-		return nil, err
-	}
-
-	switch u.Scheme {
-	case "http":
-		return net.Dial("tcp", host)
-	case "https":
-		return tls.Dial("tcp", host, &tls.Config{
-			InsecureSkipVerify: t.InsecureSkipVerify,
-		})
-	default:
-		return nil, fmt.Errorf("unknown scheme '%s'", u.Scheme)
-	}
-}
-
 func hostString(u *url.URL) (string, error) {
 	// if our Host already has a port, bail out
 	if strings.Contains(u.Host, ":") {
@@ -166,73 +120,3 @@ func hostString(u *url.URL) (string, error) {
 	}
 }
 
-func parseStatus(r *bufio.Reader) (int, error) {
-	statusLine, err := r.ReadString('\n')
-	if err != nil {
-		return 0, err
-	}
-
-	parts := strings.Split(statusLine, " ")
-	if len(parts) < 3 {
-		return 0, fmt.Errorf("'%s' is an invalid HTTP status response", strings.TrimSpace(statusLine))
-	}
-
-	status, err := strconv.Atoi(parts[1])
-	if err != nil {
-		return 0, err
-	}
-
-	return status, nil
-}
-
-func parseHeaders(r *bufio.Reader) (http.Header, error) {
-	headers := make(http.Header)
-	for {
-		line, err := r.ReadString('\n')
-
-		if err != nil {
-			return headers, nil
-		}
-
-		cleanLine := strings.TrimSpace(line)
-		if cleanLine == "" {
-			// end of headers
-			break
-		}
-
-		parts := strings.SplitN(cleanLine, ": ", 2)
-		if len(parts) == 2 {
-			headers.Add(parts[0], parts[1])
-		}
-	}
-	return headers, nil
-}
-
-func genRequest(t Target) (string, error) {
-	u, err := url.Parse(t.URL)
-	if err != nil {
-		return "", err
-	}
-
-	// allow Host header to be set via t.RequestHeaders
-	// otherwise, use the host of the URL
-	hostHeader := t.RequestHeaders["Host"]
-	if hostHeader == "" {
-		hostHeader = u.Host
-	}
-
-	// our standard request
-	req := fmt.Sprintf("GET %s HTTP/1.1\r\n", u.RequestURI())
-	req += fmt.Sprintf("Host: %s\r\n", hostHeader)
-
-	for k, v := range t.RequestHeaders {
-		if k != "Host" {
-			req += fmt.Sprintf("%s: %s\r\n", k, v)
-		}
-	}
-
-	// trailing newline
-	req += "\r\n"
-
-	return req, nil
-}
